Silence gcc > 7 type-limits warning
GCC > 7 warns that the selected type of the enum makes the comparisson
unneccesary. Since the selected enum fits in an integer, the input
parameter is casted to an integer to make the comparison valid and
asserts that the enum is within valid enum values (which we know are
monotonically growing).

Change-Id: Id81799e2ac43a83b998541ce7531d2039202cd62