Misc updates to ethosu_log.h
Replace empty log macros with calls to a new LOG_COMMON_NOP macro, with
the intention of avoiding unused variable/parameter warnings where the
variables/parameters are currently only used for logging. The if check
will normally be optimized away.

Change formatting of error messages, prepend the file name and line
instead of appending it, to allow for newlines in error messages.

Add a check for __FILE_NAME__ macro and use that instead of strrchr
function, when available.

Fix broken LOG macro

Correct copyright years

Change-Id: Ic443a6ba6d4f723c9272e605a2d975413ffe9a66
Signed-off-by: Jonny Svärd <jonny.svaerd@arm.com>